Move i18n.wrapGettext to separate module
